requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.0
    # via
    #   django
    #   django-cors-headers
    #   django-simple-history
beautifulsoup4==4.12.3
    # via -r requirements.in
certifi==2024.2.2
    # via
    #   pyproj
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==42.0.5
    # via
    #   django-rest-knox
    #   jwcrypto
    #   pyjwt
defusedxml==0.7.1
    # via python3-openid
django==5.0.3
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-cors-headers
    #   django-formtools
    #   django-oapif
    #   django-oauth-toolkit
    #   django-otp
    #   django-phonenumber-field
    #   django-rest-knox
    #   django-two-factor-auth
    #   djangorestframework
django-allauth==0.61.1
    # via -r requirements.in
django-cors-headers==4.3.1
    # via -r requirements.in
django-formtools==2.5.1
    # via django-two-factor-auth
django-migrate-sql-3==3.0.2
    # via -r requirements.in
django-oapif @ git+https://github.com/opengisch/django-oapif.git
    # via -r requirements.in
django-oauth-toolkit==2.3.0
    # via -r requirements.in
django-otp==1.3.0
    # via django-two-factor-auth
django-phonenumber-field==7.3.0
    # via django-two-factor-auth
django-rest-knox==4.2.0
    # via -r requirements.in
django-simple-history==3.5.0
    # via -r requirements.in
django-two-factor-auth[phonenumbers]==1.16.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   django-oapif
    #   django-rest-knox
    #   djangorestframework-gis
djangorestframework-gis @ git+https://github.com/openwisp/django-rest-framework-gis@4f244d5d8a7ad5b453fd04f64150818d15123e01
    # via django-oapif
gunicorn==21.2.0
    # via -r requirements.in
idna==3.6
    # via requests
jwcrypto==1.5.6
    # via django-oauth-toolkit
oauthlib==3.2.2
    # via
    #   django-oauth-toolkit
    #   requests-oauthlib
packaging==24.0
    # via gunicorn
phonenumbers==8.13.32
    # via django-two-factor-auth
psycopg==3.1.18
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via django-oapif
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via django-allauth
pypng==0.20220715.0
    # via qrcode
pyproj==3.6.1
    # via django-oapif
python3-openid==3.2.0
    # via django-allauth
pytz==2024.1
    # via djangorestframework
pyyaml==6.0.1
    # via django-oapif
qrcode==7.4.2
    # via django-two-factor-auth
requests==2.31.0
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-oauth-toolkit
    #   requests-oauthlib
requests-oauthlib==1.4.0
    # via django-allauth
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
typing-extensions==4.10.0
    # via
    #   jwcrypto
    #   psycopg
    #   qrcode
uritemplate==4.1.1
    # via django-oapif
urllib3==2.2.1
    # via requests
whitenoise==6.6.0
    # via -r requirements.in
